Mocked some mock data into a Base file since the data is only used by specific test suites
